+++ opentelemetry-sdk/src/metrics/exporter.rs
@@ -5,7 +5,7 @@ use crate::metrics::MetricResult;
 
 use crate::metrics::data::ResourceMetrics;
 
-use super::data::Temporality;
+use super::Temporality;
 
 /// Exporter handles the delivery of metric data to external receivers.
 ///